requirements.txt
numpy
Pillow
SimpleITK
nibabel
tqdm
torch
torchvision
diffusers
transformers
peft
accelerate
huggingface_hub
safetensors
